+++ src/utils/api-utils/api-constants.js
@@ -16,7 +16,6 @@ export const NUMERIC_SEARCH_TYPE = Symbol("numeric");
 export const BOOLEAN_SEARCH_TYPE = Symbol("boolean");
 export const ENUM_SEARCH_TYPE = Symbol("enum-key");
 export const FULL_TEXT_SEARCH_TYPE = Symbol("full-text");
-export const MULTI_FIELD_TEXT_SEARCH_TYPE = Symbol("multi-field-text-search");
 export const DATE_SEARCH_TYPE = Symbol("date");
 export const DATETIME_SEARCH_TYPE = Symbol("datetime");
 export const FULL_ID_SEARCH_TYPE = Symbol("full-id");