+++ src/utils/api-utils/api-util-build-search-query.spec.js
@@ -10,7 +10,6 @@ import {
   FULL_TEXT_SEARCH_TYPE,
   LIKE_ID_SEARCH_TYPE,
   LIKE_TEXT_SEARCH_TYPE,
-  MULTI_FIELD_TEXT_SEARCH_TYPE,
   NUMERIC_SEARCH_TYPE,
   STRING_END_CHAR,
   STRING_START_CHAR,
@@ -18,7 +17,7 @@ import {
   doublePipe,
   percentChar,
 } from "./api-constants";
-import { buildSearchQuery } from "./api-util";
+import { buildSearchQuery, deprecatedBuildSearchQuery } from "./api-util";
 
 moment.tz.setDefault("America/Los_Angeles");
 
@@ -36,7 +35,6 @@ describe("buildSearchQuery for full text search", () => {
           name: "column1",
           values: [""],
           type: FULL_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual("");
@@ -48,7 +46,6 @@ describe("buildSearchQuery for full text search", () => {
           name: "column1",
           values: ["abc"],
           type: FULL_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual('column1=="abc"');
@@ -63,7 +60,6 @@ describe("buildSearchQuery for like text search", () => {
           name: "column1",
           values: [""],
           type: LIKE_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual("");
@@ -75,7 +71,6 @@ describe("buildSearchQuery for like text search", () => {
           name: "column1",
           values: ["abc"],
           type: LIKE_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual(`column1=="${percentChar}abc${percentChar}"`);
@@ -87,7 +82,6 @@ describe("buildSearchQuery for like text search", () => {
           name: "column1",
           values: [`${STRING_START_CHAR}abc`],
           type: LIKE_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual(`column1=="abc${percentChar}"`);
@@ -99,7 +93,6 @@ describe("buildSearchQuery for like text search", () => {
           name: "column1",
           values: [`abc${STRING_END_CHAR}`],
           type: LIKE_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual(`column1=="${percentChar}abc"`);
@@ -111,7 +104,6 @@ describe("buildSearchQuery for like text search", () => {
           name: "column1",
           values: [`${STRING_START_CHAR}abc${STRING_END_CHAR}`],
           type: LIKE_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual('column1=="abc"');
@@ -123,7 +115,6 @@ describe("buildSearchQuery for like text search", () => {
           name: "column1",
           values: ['"abc"'],
           type: LIKE_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual('column1=="abc"');
@@ -138,7 +129,6 @@ describe("buildSearchQuery for full id search", () => {
           name: "column1",
           values: [""],
           type: FULL_ID_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual("");
@@ -150,7 +140,6 @@ describe("buildSearchQuery for full id search", () => {
           name: "column1",
           values: ["P00100-1"],
           type: FULL_ID_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual('column1=="P00100-1"||column1=="P001001"');
@@ -162,7 +151,6 @@ describe("buildSearchQuery for full id search", () => {
           name: "column1",
           values: ["A00100-1"],
           type: FULL_ID_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual('column1=="A00100-1"');
@@ -177,7 +165,6 @@ describe("buildSearchQuery for like id search", () => {
           name: "column1",
           values: [""],
           type: LIKE_ID_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual("");
@@ -189,7 +176,6 @@ describe("buildSearchQuery for like id search", () => {
           name: "column1",
           values: ["P00100-1"],
           type: LIKE_ID_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual(
@@ -217,17 +203,16 @@ describe("buildSearchQuery for like id search", () => {
           name: "column1",
           values: ["A00100-1"],
           type: LIKE_ID_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual(`column1=="${percentChar}A00100-1${percentChar}"`);
   });
 });
 
-describe("buildSearchQuery for numeric search", () => {
+describe("deprecatedBuildSearchQuery for numeric search", () => {
   it("should generate no query for empty search", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           value: undefined,
           type: NUMERIC_SEARCH_TYPE,
@@ -237,7 +222,7 @@ describe("buildSearchQuery for numeric search", () => {
   });
   it("should generate query for 0 search", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           value: 0,
           type: NUMERIC_SEARCH_TYPE,
@@ -247,7 +232,7 @@ describe("buildSearchQuery for numeric search", () => {
   });
   it("should generate query for one element", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           value: 123,
           type: NUMERIC_SEARCH_TYPE,
@@ -257,10 +242,10 @@ describe("buildSearchQuery for numeric search", () => {
   });
 });
 
-describe("buildSearchQuery for boolean search", () => {
+describe("deprecatedBuildSearchQuery for boolean search", () => {
   it("should generate no query for empty search", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           value: undefined,
           type: BOOLEAN_SEARCH_TYPE,
@@ -270,7 +255,7 @@ describe("buildSearchQuery for boolean search", () => {
   });
   it("should generate query for false search", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           value: false,
           type: BOOLEAN_SEARCH_TYPE,
@@ -280,7 +265,7 @@ describe("buildSearchQuery for boolean search", () => {
   });
   it("should generate query for one element", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           value: true,
           type: BOOLEAN_SEARCH_TYPE,
@@ -290,13 +275,13 @@ describe("buildSearchQuery for boolean search", () => {
   });
 });
 
-describe("buildSearchQuery for multi field search", () => {
+describe("deprecatedBuildSearchQuery for multi field search", () => {
   it("should generate no query for empty string search", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           value: "",
-          type: MULTI_FIELD_TEXT_SEARCH_TYPE,
+          type: LIKE_TEXT_SEARCH_TYPE,
           searchFields: ["column1", "column2"],
         }),
       ),
@@ -304,42 +289,39 @@ describe("buildSearchQuery for multi field search", () => {
   });
   it("should generate query for elements", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           value: "123",
-          type: MULTI_FIELD_TEXT_SEARCH_TYPE,
+          type: LIKE_TEXT_SEARCH_TYPE,
           searchFields: ["column1", "column2"],
         }),
       ),
-    ).toEqual(
-      `((column1=="${percentChar}123${percentChar}")${doublePipe}(column2=="${percentChar}123${percentChar}"))`,
-    );
+    ).toEqual(`(column1=="${percentChar}123${percentChar}")${doublePipe}(column2=="${percentChar}123${percentChar}")`);
   });
   it("should generate query for elements, and multiple options", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map()
           .set("column1", {
             value: "123",
-            type: MULTI_FIELD_TEXT_SEARCH_TYPE,
+            type: LIKE_TEXT_SEARCH_TYPE,
             searchFields: ["column1", "column2"],
           })
           .set("column3", {
             value: "456",
-            type: MULTI_FIELD_TEXT_SEARCH_TYPE,
-            searchFields: ["column3"],
+            type: LIKE_TEXT_SEARCH_TYPE,
           }),
       ),
     ).toEqual(
-      `((column1=="${percentChar}123${percentChar}")${doublePipe}(column2=="${percentChar}123${percentChar}"))${doubleAmpersand}((column3=="${percentChar}456${percentChar}"))`,
+      `((column1=="${percentChar}123${percentChar}")${doublePipe}(column2=="${percentChar}123${percentChar}"))${doubleAmpersand}(column3=="${percentChar}456${percentChar}")`,
     );
   });
 });
 
-describe("buildSearchQuery for multi value search", () => {
+describe("deprecatedBuildSearchQuery for multi value search", () => {
   it("should generate no query for empty array", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           values: [],
           type: FULL_TEXT_SEARCH_TYPE,
@@ -349,7 +331,7 @@ describe("buildSearchQuery for multi value search", () => {
   });
   it("should generate query for Like text search elements", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           type: LIKE_TEXT_SEARCH_TYPE,
           values: ["123", "345"],
@@ -359,7 +341,7 @@ describe("buildSearchQuery for multi value search", () => {
   });
   it("should generate query for Numeric text search elements", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           type: NUMERIC_SEARCH_TYPE,
           values: [123, 345.6],
@@ -369,7 +351,7 @@ describe("buildSearchQuery for multi value search", () => {
   });
   it("should generate query for Boolean text search elements", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           type: BOOLEAN_SEARCH_TYPE,
           values: [true, false], // Technically, this is silly...
@@ -379,7 +361,7 @@ describe("buildSearchQuery for multi value search", () => {
   });
   it("should generate query for Full text search elements", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("column1", {
           type: FULL_TEXT_SEARCH_TYPE,
           values: ["123", "345"],
@@ -389,7 +371,7 @@ describe("buildSearchQuery for multi value search", () => {
   });
 });
 
-describe("buildSearchQuery for datetime search", () => {
+describe("deprecatedBuildSearchQuery for datetime search", () => {
   const date = "2017-04-20T16:20:00.000Z";
   const startOfDay = moment(date)
     .startOf("day")
@@ -399,7 +381,7 @@ describe("buildSearchQuery for datetime search", () => {
     .toISOString();
   it("should generate query for start date", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("date", {
           values: [`${date}`],
           type: DATETIME_SEARCH_TYPE,
@@ -409,7 +391,7 @@ describe("buildSearchQuery for datetime search", () => {
   });
   it("should generate query for start date to", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("date", {
           values: [`${date} to`],
           type: DATETIME_SEARCH_TYPE,
@@ -419,7 +401,7 @@ describe("buildSearchQuery for datetime search", () => {
   });
   it("should generate query for end date with no string startDate", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("date", {
           values: [`to ${date}`],
           type: DATETIME_SEARCH_TYPE,
@@ -433,7 +415,7 @@ describe("buildSearchQuery for datetime search", () => {
     .toISOString();
   it("should generate query for start and end date", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("date", {
           values: [`${newStartDate} to ${date}`],
           type: DATETIME_SEARCH_TYPE,
@@ -443,7 +425,7 @@ describe("buildSearchQuery for datetime search", () => {
   });
   it("should generate query for reversed start and end date", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("date", {
           values: [`${date} to ${newStartDate}`],
           type: DATETIME_SEARCH_TYPE,
@@ -453,7 +435,7 @@ describe("buildSearchQuery for datetime search", () => {
   });
   it("should generate query for no dates", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("date", {
           values: [],
           type: DATETIME_SEARCH_TYPE,
@@ -463,11 +445,11 @@ describe("buildSearchQuery for datetime search", () => {
   });
 });
 
-describe("buildSearchQuery for date search", () => {
+describe("deprecatedBuildSearchQuery for date search", () => {
   const date = "2017-04-20";
   it("should generate query for start date", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("date", {
           values: [`${date}`],
           type: DATE_SEARCH_TYPE,
@@ -477,7 +459,7 @@ describe("buildSearchQuery for date search", () => {
   });
   it("should generate query for start date to", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("date", {
           values: [`${date} to`],
           type: DATE_SEARCH_TYPE,
@@ -487,7 +469,7 @@ describe("buildSearchQuery for date search", () => {
   });
   it("should generate query for end date with no string startDate", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("date", {
           values: [`to ${date}`],
           type: DATE_SEARCH_TYPE,
@@ -498,7 +480,7 @@ describe("buildSearchQuery for date search", () => {
   const newStartDate = "2016-04-20";
   it("should generate query for start and end date", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("date", {
           values: [`${newStartDate} - ${date}`],
           type: DATE_SEARCH_TYPE,
@@ -508,7 +490,7 @@ describe("buildSearchQuery for date search", () => {
   });
   it("should generate query for reversed start and end date", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("date", {
           values: [`${date} - ${newStartDate}`],
           type: DATE_SEARCH_TYPE,
@@ -518,7 +500,7 @@ describe("buildSearchQuery for date search", () => {
   });
   it("should generate query for no dates", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map().set("date", {
           values: [],
           type: DATE_SEARCH_TYPE,
@@ -528,10 +510,10 @@ describe("buildSearchQuery for date search", () => {
   });
 });
 
-describe("buildSearchQuery for several search items", () => {
+describe("deprecatedBuildSearchQuery for several search items", () => {
   it("should generate query for several search elements", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map()
           .set("column1", {
             value: 123,
@@ -552,7 +534,7 @@ describe("buildSearchQuery for several search items", () => {
   });
   it("should generate query for several search elements when one is empty string", () => {
     expect(
-      buildSearchQuery(
+      deprecatedBuildSearchQuery(
         new Map()
           .set("column1", {
             value: 123,
@@ -571,17 +553,18 @@ describe("buildSearchQuery for several search items", () => {
   });
   it("should generate query for several search elements when the last string is empty string (See T1661)", () => {
     expect(
-      buildSearchQuery(
-        new Map()
-          .set("column1", {
-            value: 123,
-            type: NUMERIC_SEARCH_TYPE,
-          })
-          .set("column2", {
-            value: "",
-            type: FULL_TEXT_SEARCH_TYPE,
-          }),
-      ),
+      buildSearchQuery([
+        {
+          searchFields: ["column1"],
+          values: ["123"],
+          type: NUMERIC_SEARCH_TYPE,
+        },
+        {
+          searchFields: ["column2"],
+          values: [""],
+          type: FULL_TEXT_SEARCH_TYPE,
+        },
+      ]),
     ).toEqual("(column1==123)");
   });
 });
@@ -608,7 +591,6 @@ describe("buildSearchQuery is invalid", () => {
           name: "column1",
           values: [],
           type: DATE_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual("");
@@ -626,7 +608,6 @@ describe("buildSearchQuery for uri encoding", () => {
           name: "column1",
           values: ["abc & 123", "xyz"],
           type: FULL_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
         },
       ]),
     ).toEqual(`column1=="abc%20%26%20123"${doublePipe}column1=="xyz"`);
@@ -641,7 +622,6 @@ describe("buildSearchQuery for searchOperator", () => {
           name: "column1",
           values: ["abc & 123", "xyz"],
           type: FULL_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
           searchOperator: "!=",
         },
       ]),
@@ -651,7 +631,7 @@ describe("buildSearchQuery for searchOperator", () => {
     expect(
       buildSearchQuery([
         {
-          name: "column1",
+          name: "Multiple Columns",
           values: ["abc & 123", "xyz"],
           type: FULL_TEXT_SEARCH_TYPE,
           searchFields: ["column1", "column2"],
@@ -672,7 +652,6 @@ describe("buildSearchQuery for includeNulls", () => {
           name: "column1",
           values: [],
           type: FULL_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
           includeNulls: true,
         },
       ]),
@@ -685,7 +664,6 @@ describe("buildSearchQuery for includeNulls", () => {
           name: "column1",
           values: ["abc"],
           type: FULL_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
           includeNulls: true,
         },
       ]),
@@ -698,7 +676,6 @@ describe("buildSearchQuery for includeNulls", () => {
           name: "column1",
           values: ["abc & 123", "xyz"],
           type: FULL_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
           includeNulls: true,
         },
       ]),
@@ -708,7 +685,7 @@ describe("buildSearchQuery for includeNulls", () => {
     expect(
       buildSearchQuery([
         {
-          name: "column1",
+          name: "Multiple Columns",
           values: ["abc & 123", "xyz"],
           type: FULL_TEXT_SEARCH_TYPE,
           searchFields: ["column1", "column2"],
@@ -726,21 +703,18 @@ describe("buildSearchQuery for includeNulls", () => {
           name: "column1",
           values: ["xyz"],
           type: FULL_TEXT_SEARCH_TYPE,
-          searchFields: ["column1"],
           includeNulls: true,
         },
         {
-          name: "columnc",
+          name: "column2",
           values: ["xyz"],
           type: FULL_TEXT_SEARCH_TYPE,
-          searchFields: ["column2"],
           includeNulls: false,
         },
         {
-          name: "columnd",
+          name: "column3",
           values: ["abc"],
           type: FULL_TEXT_SEARCH_TYPE,
-          searchFields: ["column3"],
           includeNulls: true,
         },
       ]),